+++ app/src/main/java/com/arctouch/easybus/search/SearchActivity.java
@@ -1,17 +1,29 @@
 package com.arctouch.easybus.search;
 
+import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v7.app.AppCompatActivity;
 
-import com.arctouch.easybus.util.SingleFragmentActivity;
+import com.arctouch.easybus.R;
 
 /**
  * Activity that serves as a simple holder for the SearchFragment.
  */
-public class SearchActivity extends SingleFragmentActivity {
+public class SearchActivity extends AppCompatActivity {
 
     @Override
-    public Fragment createFragment() {
-        return new SearchFragment();
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_fragment);
+
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        Fragment fragment = fragmentManager.findFragmentById(R.id.fragment_container);
+        if (fragment == null) {
+            fragmentManager.beginTransaction()
+                    .add(R.id.fragment_container, new SearchFragment())
+                    .commit();
+        }
     }
 
 }
