--- app/src/main/java/com/arctouch/easybus/util/SingleFragmentActivity.java
@@ -1,28 +1,0 @@
-package com.arctouch.easybus.util;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
-
-import com.arctouch.easybus.R;
-
-public abstract class SingleFragmentActivity extends AppCompatActivity {
-
-    protected abstract Fragment createFragment();
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_fragment);
-
-        FragmentManager fragmentManager = getSupportFragmentManager();
-        Fragment fragment = fragmentManager.findFragmentById(R.id.fragment_container);
-        if (fragment == null) {
-            fragmentManager.beginTransaction()
-                    .add(R.id.fragment_container, createFragment())
-                    .commit();
-        }
-    }
-
-}
