+++ app/src/main/res/values/strings.xml
@@ -10,7 +10,7 @@
     <string name="schedule_menu_item_title">Schedule</string>
     <string name="route_progress_dialog_message">Loading…</string>
 
-    <string name="fetching_address">Fetching address...</string>
+    <string name="fetching_address">Fetching address…</string>
     <string name="address_not_found">No address found</string>
     <string name="on_connection_failed_message">Connection with Google Maps has failed.</string>
     <string name="on_connection_suspended_message">Connection with Google Maps has been suspended.</string>
